+++ src/sctoolbox/plotting/marker_genes.py
@@ -16,7 +16,6 @@ from matplotlib.patches import Patch
 
 from beartype.typing import Optional, Tuple, Literal, Any
 from beartype import beartype
-from numpy.typing import NDArray
 
 # sctoolbox functions
 import sctoolbox.utils as utils
@@ -693,7 +692,7 @@ def plot_gene_correlation(adata: sc.AnnData,
                           ncols: int = 3,
                           figsize: Optional[Tuple[int | float, int | float]] = None,
                           save: Optional[str] = None,
-                          **kwargs: Any) -> NDArray[Axes]:
+                          **kwargs: Any) -> np.ndarray[tuple[int, ...], np.dtype[Axes]]:
     """
     Plot the gene expression of one reference gene against the expression of a set of genes.
 
@@ -716,7 +715,7 @@ def plot_gene_correlation(adata: sc.AnnData,
 
     Returns
     -------
-    NDArray[Axes]
+    np.ndarray[tuple[int, ...], np.dtype[Axes]]
         List containing all axis objects.
 
     Examples